requirements.txt
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
colorama==0.4.6
fastapi==0.115.6
h11==0.14.0
httptools==0.6.4
idna==3.10
outcome==1.3.0.post0
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
PyMySQL==1.1.1
PySocks==1.7.1
python-dotenv==1.0.1
PyYAML==6.0.2
selenium==4.27.1
selenium-stealth==1.0.6
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.41.3
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.33.0
watchfiles==1.0.3
websocket-client==1.8.0
websockets==14.1
wsproto==1.2.0
